requirements.txt
numpy~=1.19.3
opencv-python~=4.5.1.48
Pillow~=7.1.1
PyAutoGUI~=0.9.50